fix: e-patches crashes from missing config setting
Looks like the config object structure was refactored a little and this
didn't get the memo. childProcess was also missing, which is concerning
since it implies the previous commit couldn't've worked regardless of
config.

It might be worthwhile to give e-patches.js a little love in a followup
PR in the slow-tests section. Since it's unusable right now, I'm PR'ing
this separately.