fix: don't pollute stdout with passive update logs
Bug description: scripts that rely on build-tools will fail occasionally
if build-tools hasn't had a recent update check. For example some of my
shell scripts have calls along the lines of `FOO="$(e show exec)"` and
fail when FOO has "Checking for build-tools updates" instead of a path.

When the update script is run passively, its stdout is redirected to
stderr: it's still useful there as a user-visible message, but doesn't
interfere with expected values in stdout.