fix: use execFileSync when using a cmd with a path (#178)
Previously execSync was used, taking the full command line as a single
string. This would be misparsed if the path had spaces in it. For example
"C:\Users\My Name\e\third_party\depot_tools\gn" would see "C:\Users\My"
as the command.

NB: the new code in e-build.js is cleaner, but it wasn't clear (to me,
anyway) that the changed handling of escapes would work. FWIW I tested
this change on both Win and POSIX.