feat: auto-update support of 'main' branch
We currently can't rename build-tools' primary branch from `master` to `main`
because it will break auto-updating.

This PR checks to for `main` at git@github.com:electron/build-tools.git and,
if found, recommends the user to switch to _that_ branch instead of `master`.

So we can merge this, then wait a while for users to get auto-updated with
this code, and then make the switch.